api/requirements.txt
fastapi[all]
websockets
pyserial
pypubsub